auth

+++ main.py
@@ -77,10 +77,21 @@ def products():
 
 #ดึงproductตามid
 @app.route('/api/products/<id>' ,methods=['GET'])
+@jwt_required()
 def products_id_losts(id):
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         sql = "SELECT * FROM products  WHERE products.id_lots = %s;"
         val = (f"{id}",)
         mycursor.execute(sql,val)
@@ -93,20 +104,27 @@ def products_id_losts(id):
 
 #insert product
 @app.route('/api/products', methods=['POST'])
+@jwt_required()
 def products_insert():
     try:
         mydb = mysql.connector.connect(host=host, user=user, password=password, database=database)
         data = request.get_json()
         mycursor = mydb.cursor(dictionary=True)
-        query = "SELECT * FROM user WHERE user.name = %s"
-        valquery = (data['user'],)
-        mycursor.execute(query,valquery)
-        result = mycursor.fetchall()
-        print(result[0]['id'])
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
+        print(user_result['id'])
         sql = """INSERT INTO products (id_lots, id_user, BreakClean, CompleteSeeds, Dust, MoldSpores, broken, fullbrokenseeds, path)
                  VALUES (%s, %s, %s, %s, %s, %s, %s, %s, %s)"""
         val = (
-             data['id'],result[0]['id'], data['BreakClean'], data['CompleteSeeds'], data['Dust'],
+             data['id'],user_result['id'], data['BreakClean'], data['CompleteSeeds'], data['Dust'],
             data['MoldSpores'], data['broken'], data['fullbrokenseeds'], data['path']
         )
         mycursor.execute(sql, val)
@@ -121,9 +139,20 @@ def products_insert():
 
 #update product
 @app.route('/api/products', methods=['PUT'])
+@jwt_required()
 def products_update():
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         data = request.get_json()
         mycursor = mydb.cursor(dictionary=True)
         sql = "UPDATE `products` SET `id_lots`=%s WHERE products.id = %s;"
@@ -138,10 +167,21 @@ def products_update():
 
 #delete product
 @app.route('/api/products/<id>', methods=['DELETE'])
+@jwt_required()
 def products_delete(id):
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         sql = "DELETE FROM `products` WHERE products.id = %s;"
         val = (f"{id}",)
         mycursor.execute(sql, val)  
@@ -160,6 +200,7 @@ def products_delete(id):
 
 #ดึงข้อมูลเป็นหน้าๆ
 @app.route('/api/lots/<page>', methods=['GET'])
+@jwt_required()
 def lots_page(page):
     try:
         page_value = int(page)
@@ -185,6 +226,16 @@ def lots_page(page):
         values = (max_value, min_value)
         with mysql.connector.connect(host=host, user=user, password=password, database=database) as mydb:
             with mydb.cursor(dictionary=True) as mycursor:
+                # Get current user from JWT
+                current_user = get_jwt_identity()
+                # Check if the user exists
+                sql = "SELECT * FROM user WHERE name = %s"
+                val = (current_user,)
+                mycursor.execute(sql, val)
+                user_result = mycursor.fetchone()
+                # If user does not exist, return an error response
+                if not user_result:
+                    return make_response(jsonify({"msg": "Token is bad"}), 404)
                 mycursor.execute(sql, values)
                 myresult = mycursor.fetchall()
     except Error as e:
@@ -195,10 +246,21 @@ def lots_page(page):
 
 #ดึงจำนวนหน้า
 @app.route('/api/lots/sum',methods=['GET'])
+@jwt_required()
 def lots_sum():
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         sql = """SELECT COUNT(lots.id) as sum
             FROM lots INNER JOIN (SELECT p.id_lots AS max_id, p.path FROM products AS p
             JOIN (SELECT MAX(id) AS id, id_lots FROM products GROUP BY id_lots) AS p_max 
@@ -218,10 +280,21 @@ def lots_sum():
 
 #ดึงจำนวนหน้าที่ค้นหา
 @app.route('/api/lots/search/sum',methods=['POST'])
+@jwt_required()
 def lots_like_sum():
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         data = request.get_json()
         id = data.get('id')
         sql = """SELECT COUNT(lots.id) as sum
@@ -245,10 +318,21 @@ def lots_like_sum():
 
 #ดึงหน้าที่ค้นหา
 @app.route('/api/lots/search',methods=['POST'])
+@jwt_required()
 def lots_like_id():
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         data = request.get_json()
         page = data.get('page')
         id = data.get('id')
@@ -279,10 +363,21 @@ def lots_like_id():
 
 #ดึงล็อตตามid
 @app.route('/api/lotId/<id>', methods=['GET'])
+@jwt_required()
 def lots_id(id):
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         sql = ("""SELECT * FROM lots 
                 INNER JOIN products ON products.id_lots = lots.id
                 WHERE lots.id = %s;""")
@@ -297,10 +392,21 @@ def lots_id(id):
 
 #ดึงล็อตตามid
 @app.route('/api/lotstatusId/<id>', methods=['GET'])
+@jwt_required()
 def lots_idstatus(id):
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         sql = ("""SELECT * FROM lots LEFT JOIN
                 (SELECT status.id_lots as id,status.status as status FROM status INNER JOIN 
                 (SELECT max(status.id) as id FROM status GROUP BY status.id_lots) 
@@ -318,10 +424,23 @@ def lots_idstatus(id):
 
 #ดึงล็อตทั้งหมด
 @app.route('/api/lots', methods=['GET'])
+@jwt_required()
 def lots():
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
+        
         sql = "SELECT * FROM lots;"
         mycursor.execute(sql)
         myresult = mycursor.fetchall()
@@ -333,10 +452,21 @@ def lots():
 
 #สร้างล็อต
 @app.route('/api/lots', methods=['POST'])
+@jwt_required()
 def lots_insert():
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         data = request.get_json()
         now = datetime.now()
         current_time = now.strftime("%Y-%m-%d")
@@ -353,10 +483,21 @@ def lots_insert():
 
 #ลบล็อต
 @app.route('/api/lots/<id>', methods=['DELETE'])
+@jwt_required()
 def lots_delete(id):
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         sql = ("DELETE FROM lots WHERE lots.id = %s;")
         val =(f"{id}",)
         mycursor.execute(sql, val)
@@ -369,10 +510,21 @@ def lots_delete(id):
 
 #ดึงล็อตทั้งหมด
 @app.route('/api/graph/<id>', methods=['GET'])
+@jwt_required()
 def lots_productgraph(id):
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         val =(f"{id}",)
         sql = """SELECT (SUM(p.BreakClean)/ (SUM(p.BreakClean)+SUM(p.CompleteSeeds)+SUM(p.Dust)+SUM(p.MoldSpores)+SUM(p.broken)+SUM(p.fullbrokenseeds)))*100 as BreakClean,
             (SUM(p.CompleteSeeds)/ (SUM(p.BreakClean)+SUM(p.CompleteSeeds)+SUM(p.Dust)+SUM(p.MoldSpores)+SUM(p.broken)+SUM(p.fullbrokenseeds)))*100 as CompleteSeeds,
@@ -406,6 +558,7 @@ def lots_productgraph(id):
 ######################################################################
 # API Login
 @app.route('/api/login', methods=['POST'])
+@jwt_required()
 def login():
     data = request.get_json()
     if data:
@@ -442,10 +595,21 @@ def login():
 #API Status
 #get status
 @app.route('/api/status/<id>',methods=['GET'])
+@jwt_required()
 def status_id(id):
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         sql = "SELECT * FROM status WHERE status.id_lots = %s ORDER BY status.id DESC LIMIT 1;"
         val = (f"{id}",)
         mycursor.execute(sql,val)
@@ -458,10 +622,21 @@ def status_id(id):
 
 #insert status
 @app.route('/api/status',methods=['POST'])
+@jwt_required()
 def status_insert():
     try:
         mydb = mysql.connector.connect(host=host,user=user,password=password,database=database)
         mycursor = mydb.cursor(dictionary=True)
+        # Get current user from JWT
+        current_user = get_jwt_identity()
+        # Check if the user exists
+        sql = "SELECT * FROM user WHERE name = %s"
+        val = (current_user,)
+        mycursor.execute(sql, val)
+        user_result = mycursor.fetchone()
+        # If user does not exist, return an error response
+        if not user_result:
+            return make_response(jsonify({"msg": "Token is bad"}), 404)
         data = request.get_json()
         sql = """INSERT INTO `status`(`id_lots`, `id_user`, `status`, `date`) VALUES 
         (%s,%s,%s,%s)"""
@@ -531,7 +706,24 @@ def handle_request_video():
 
 # ลบรูปที่ถ่าย
 @app.route('/delete_capture', methods=['POST'])
+@jwt_required()
 def delete_capture():
+    # Connect to the database
+    mydb = mysql.connector.connect(host=host, user=user, password=password, database=database)
+    mycursor = mydb.cursor(dictionary=True)
+        
+    # Get current user from JWT
+    current_user = get_jwt_identity()
+        
+    # Check if the user exists
+    sql = "SELECT * FROM user WHERE name = %s"
+    val = (current_user,)
+    mycursor.execute(sql, val)
+    user_result = mycursor.fetchone()
+        
+    # If user does not exist, return an error response
+    if not user_result:
+        return make_response(jsonify({"msg": "Token is bad"}), 404)
     data = request.get_json()
     filename = data['filename']
     if os.path.exists(filename):
